pyproject.toml
[project]
name = "ngumpulin-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "gliner>=0.2.16",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.44",
    "langchain-pinecone>=0.2.3",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
    "supabase>=2.13.0",
    "uvicorn>=0.34.0",
]
